+++ src/Yoti.Auth/Anchors/AnchorCertificateParser.cs
@@ -9,39 +9,46 @@ using Org.BouncyCastle.Asn1;
 
 namespace Yoti.Auth.Anchors
 {
-    public static class AnchorCertificateParser
+    internal static class AnchorCertificateParser
     {
-        public static AnchorVerifierSourceData GetTypesFromAnchor(Anchor anchor)
+        public static AnchorVerifierSourceData GetTypesFromAnchor(Anchor anchor, AnchorType type)
         {
             var types = new HashSet<string>();
             AnchorType anchorType = AnchorType.Unknown;
 
             foreach (ByteString byteString in anchor.OriginServerCerts)
             {
-                var extensions = new List<string>();
                 X509Certificate2 certificate = new X509Certificate2(byteString.ToByteArray());
-                var anchorEnum = typeof(AnchorType);
+                var extensions = new List<string>();
 
-                foreach (AnchorType type in Enum.GetValues(anchorEnum))
+                foreach (AnchorType enumType in Enum.GetValues(typeof(AnchorType)))
                 {
-                    var name = Enum.GetName(anchorEnum, type);
-                    string extensionOid = anchorEnum.GetRuntimeField(name)
-                        .GetCustomAttributes(inherit: false)
-                        .OfType<ExtensionOidAttribute>()
-                        .Single().ExtensionOid;
-
-                    extensions = GetListOfStringsFromExtension(certificate, extensionOid);
+                    extensions = GetExtensions(enumType, certificate, extensions);
 
                     if (extensions.Count() > 0)
                     {
-                        anchorType = type;
+                        anchorType = enumType;
                         break;
                     }
                 }
+
                 types.UnionWith(extensions);
             }
 
             return new AnchorVerifierSourceData(types, anchorType);
+        }
+
+        private static List<string> GetExtensions(AnchorType anchorType, X509Certificate2 certificate, List<string> extensions)
+        {
+            var anchorEnum = typeof(AnchorType);
+
+            var name = Enum.GetName(anchorEnum, anchorType);
+            string extensionOid = anchorEnum.GetRuntimeField(name)
+                .GetCustomAttributes(inherit: false)
+                .OfType<ExtensionOidAttribute>()
+                .Single().ExtensionOid;
+
+            return GetListOfStringsFromExtension(certificate, extensionOid);
         }
 
         private static List<string> GetListOfStringsFromExtension(X509Certificate2 certificate, string extensionOid)