+++ src/Yoti.Auth/YotiImageAttribute.cs
@@ -1,19 +1,12 @@
-﻿using System.Collections.Generic;
+﻿using AttrpubapiV1;
+using Google.Protobuf.Collections;
 using static Yoti.Auth.YotiAttributeValue;
 
 namespace Yoti.Auth
 {
     public class YotiImageAttribute<T> : YotiAttribute<T> where T : Image
     {
-        public YotiImageAttribute(string name, YotiAttributeValue value) : base(name, value)
-        {
-        }
-
-        public YotiImageAttribute(string name, YotiAttributeValue value, HashSet<string> sources) : base(name, value, sources)
-        {
-        }
-
-        public YotiImageAttribute(string name, YotiAttributeValue value, HashSet<string> sources, HashSet<string> verifiers) : base(name, value, sources, verifiers)
+        public YotiImageAttribute(string name, YotiAttributeValue value, RepeatedField<Anchor> anchors) : base(name, value, anchors)
         {
         }