+++ test/Yoti.Auth.Tests/YotiAttribute_Tests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using Google.Protobuf.Collections;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using static Yoti.Auth.YotiAttributeValue;
 
@@ -39,7 +40,7 @@ namespace Yoti.Auth.Tests
             byte[] imageBytes = Conversion.UtfToBytes("ImageValue");
 
             var yotiAttributeValue = new YotiAttributeValue(TypeEnum.Jpeg, imageBytes);
-            var yotiAttribute = new YotiImageAttribute<Image>("selfie", yotiAttributeValue);
+            var yotiAttribute = new YotiImageAttribute<Image>("selfie", yotiAttributeValue, anchors: null);
 
             var expectedImage = new Image
             {
@@ -59,38 +60,11 @@ namespace Yoti.Auth.Tests
             var yotiAttribute = new YotiImageAttribute<Image>(
                 name: "selfie",
                 value: null,
-                sources: new HashSet<string>());
+                anchors: null);
 
             string defaultValue = "default";
 
             Assert.AreEqual(defaultValue, yotiAttribute.GetValueOrDefault(defaultValue));
-        }
-
-        [TestMethod]
-        public void YotiAttribute_GetSources()
-        {
-            var sources = new HashSet<string> { "a", "b", "c" };
-
-            var yotiAttribute = new YotiImageAttribute<Image>(
-                name: "selfie",
-                value: null,
-                sources: sources);
-
-            Assert.AreEqual(sources, yotiAttribute.GetSources());
-        }
-
-        [TestMethod]
-        public void YotiAttribute_GetVerifiers()
-        {
-            var verifiers = new HashSet<string> { "x", "y", "z" };
-
-            var yotiAttribute = new YotiImageAttribute<Image>(
-                name: "selfie",
-                value: null,
-                sources: new HashSet<string> { "a", "b", "c" },
-                verifiers: verifiers);
-
-            Assert.AreEqual(verifiers, yotiAttribute.GetVerifiers());
         }
 
         [TestMethod]
@@ -100,7 +74,7 @@ namespace Yoti.Auth.Tests
             var yotiAttributeValue = new YotiAttributeValue(
                 TypeEnum.Jpeg,
                 jpegBytes);
-            var yotiAttribute = new YotiImageAttribute<Image>("selfie", yotiAttributeValue);
+            var yotiAttribute = new YotiImageAttribute<Image>("selfie", yotiAttributeValue, anchors: null);
 
             string expectedString = String.Format("data:image/jpeg;base64,{0}", Conversion.BytesToBase64(jpegBytes));
 
@@ -114,7 +88,7 @@ namespace Yoti.Auth.Tests
             var yotiAttributeValue = new YotiAttributeValue(
                 TypeEnum.Png,
                 pngBytes);
-            var yotiAttribute = new YotiImageAttribute<Image>("selfie", yotiAttributeValue);
+            var yotiAttribute = new YotiImageAttribute<Image>("selfie", yotiAttributeValue, anchors: null);
 
             string expectedString = String.Format("data:image/png;base64,{0}", Conversion.BytesToBase64(pngBytes));